requirements.txt
pytube
ffmpeg-python
bing-image-downloader
moviepy